[feat] #231 elasticsearch 리뷰개수순 정렬 추가

+++ src/main/java/com/dmarket/dto/common/ProductCommonDto.java
@@ -130,6 +130,6 @@ public class ProductCommonDto {
         private Integer productSalePrice;
         private Integer productDiscountRate;
         private Float productRating;
-        private Long productReviewCount;
+        private int productReviewCount;
     }
 }
